RSA changed to return short int vector in encrypt

+++ Wav_project/EncryptionAlgorithms.h
@@ -21,7 +21,7 @@ namespace EncryptionAlgorithms
 
 	std::vector<int> xor (const std::vector<int>& inputVector, const std::vector<int> keyVector);
 
-	std::vector<int> encryptRsa8 (const std::vector<short int>& inputData, const short int & e, const short int & n);
-	std::vector<short int> decryptRsa8(const std::vector<int>& inputData, const short int & e, const short int & n);
+	std::vector<short int> encryptRsa8 (const std::vector<short int>& inputData, const short int & e, const short int & n);
+	std::vector<short int> decryptRsa8(const std::vector<short int>& inputData, const short int & e, const short int & n);
 };
